requirements.txt
django >= 2.1
        pyotp
        python-u2flib-server
        ua-parser
        user-agents
        python-jose
        fido2 == 0.7

